requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
requests
PyMuPDF
numpy
faiss-cpu
google-genai
python-multipart
